webapi/webapi-webspeech-w3c-tests/webspeech/speechsynthesisutterance-basics.html

<!DOCTYPE html>
<!--
Copyright (c) 2014 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Wu. Xiaojun <xiaojunx.a.wu@intel.com>

-->

<meta charset="utf-8">
<title>Web Speech API Test: the basics of the SpeechRecognitionEvent interface</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://dvcs.w3.org/hg/speech-api/raw-file/tip/speechapi.html">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="log"></div>
<script src="../resources/vendor-prefix.js" data-prefixed-objects='[{"ancestors":["window"], "name":"SpeechSynthesisUtterance"}]'></script>
<script>

setup(function () {
  window.ssu = new SpeechSynthesisUtterance();
});

test(function () {
  assert_true("SpeechSynthesisUtterance" in window);
}, "Existence of SpeechSynthesisUtterance interface");

test(function () {
  assert_not_equals(SpeechSynthesisUtterance, null);
}, "Value of SpeechSynthesisUtterance object");

[
  "text", "lang", "voiceURI", "volume", "rate", "pitch",
  "onstart", "onend", "onerror", "onpause", "onresume", "onmark", "onboundary"
].forEach(function(attr) {
  test(function () {
    assert_true("text" in ssu);
  }, "Existence of " + attr + " attribute");
});

[
  {name: "text", normalV:"Sample text"},
  {name: "lang", normalV:"en"},
  {name: "voiceURI", normalV:"http://example.com"},
  {name:"volume", defaultV:1, normalV:0.5, minV:0, maxV:1},
  {name:"rate", defaultV:1, normalV:2, minV:0.1, maxV:10},
  {name:"pitch", defaultV:1, normalV:1.5, minV:0, maxV:2}
].forEach(function(attr) {
  if (attr["defaultV"] !== undefined)
    test(function(){
      assert_equals(ssu[attr["name"]], attr["defaultV"], "The default value of " + attr["name"] + " attribute should be "+ attr["defaultV"] +".");
    }, "Default value of " + attr["name"] + " attribute");

  test(function(){
    ssu[attr["name"]] = attr["normalV"];
    assert_equals(ssu[attr["name"]], attr["normalV"], "The value of " + attr["name"] + " attribute should be "+ attr["normalV"] +".");
  }, "Set value for " + attr["name"] + " attribute");

  if (attr["minV"] !== undefined)
    test(function(){
      ssu[attr["name"]] = attr["minV"] - 0.1;
      assert_equals(ssu[attr["name"]], attr["minV"], "The value of " + attr["name"] + " attribute should be converted to min value(" + attr["minV"] + ").");
    }, "Min value of " + attr["name"] + " attribute");
  if (attr["maxV"] !== undefined)
    test(function() {
      ssu[attr["name"]] = attr["maxV"] + 0.1;
      assert_equals(ssu[attr["name"]], attr["maxV"], "The value of " + attr["name"] + " attribute should be converted to max value(" + attr["maxV"] + ").");
    }, "Max value of " + attr["name"] + " attribute");
});

</script>
